+++ src/App.css
@@ -1,145 +1,106 @@
-body{
-  background-color: hsl(220, 13%, 20%);
+.appmain{
+  display: grid;
+height: 100%;
+width: 100%;
+display: grid;
+grid-template-rows: 75px 600px 125px;
+grid-template-columns: 100%;
+margin: auto;
+background-color: black;
 }
 
-h2 {
+.main-header{
+  display: grid;
+  grid-template-columns: 25% 75%;
+
+}
+
+.main-body{
   display: flex;
+  justify-content: center;
+  align-items: center;
+  color:white;
+
 }
 
+.main-footer{
+  display: grid;
+  justify-content: center;
 
-h2 * {
-  margin: 0 5px;
 }
 
+li{
+  display:inline-block;
+  margin: auto;
+  font-size: 20px;
+  padding: 15px;
+}
+li a{
+  display: block;
+  padding: 8px;
+  color: white;
+}
+button{
+  display: block;
+  text-align: center;
+  color: white;
+  font-size: 20px;
+  background-color: transparent;
+    background-repeat: no-repeat;
+    border: none;
+    cursor: pointer;
+    overflow: hidden;
+    outline: none;
+}
+
+a:link {
+  text-decoration: none;
+}
+a:visited {
+  text-decoration: none;
+}
 
 .App-logo {
-  height: 4vmin;
+  height: 8vmin;
   pointer-events: none;
 }
-.Function-button{
-  align-items: right;
-}
-.container {
-  
-  background-color: hsl(220, 13%, 10%);
-  min-height: 2em;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  font-size: calc(8px + 1vmin);
-  color: white;
-  
-}
+.macMockup {
+  height: 50vmin;
+  pointer-events: none;
+  float: left;
 
-.content{
-  margin-right: auto;
-  margin-left: 20em;
-  color: white;
-  
 }
-
-.rows{
-  margin: 10px 0px
+.nav-logo{
+  float: left;
+  background-color: black;
+  padding-left: 15%;
+  padding-top: 5%;
+  color:white;
 }
-
 .nav-links{
-  margin-right: 15em;
-  margin-left: auto;
+  float:right;
+  padding-right: 10%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: 20px;
+  border-color: blue; 
+}
+#getaccount{
+  background-color: red;
+}
+#getaccount:hover{
+  background-color: gray;
 }
 
-.nav-links * {
-  list-style-type: none;
-  padding: 10px;
-}
-
-a {
-  text-decoration: none;
-  color: #e6e6e6;
-}
-
-a:visited {
-  color: #e6e6e6;
-}
-
-#image{
-  margin-right: auto;
-  margin-left: 15em;
-}
-
-.modal-cover {
+.modal {
+  display: none;
   position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  z-index: 10;
-  transform: translateZ(0);
-  background-color: rgba(0, 0, 0, 0.8);
+  z-index: 1;
+  background-color: rgba(0, 0, 0, 0.25);
 }
 
-.modal-area {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  padding: 2.5em 1.5em 1.5em 1.5em;
-  background-color: #ffffff;
-  box-shadow: 0 0 10px 3px rgba(0, 0, 0, 0.1);
-  overflow-y: auto;
-  -webkit-overflow-scrolling: touch;
-}
-
-@media screen and (min-width: 500px) {
-  /* Center the Modal! */
-  .modal-area {
-    left: 50%;
-    top: 50%;
-    height: auto;
-    transform: translate(-50%, -50%);
-    max-width: 30em;
-    max-height: calc(100% - 1em);
-  }
-}
-
-._modal-close {
+.modal_content {
+  background-color: white;
   position: absolute;
-  top: 0;
-  right: 0;
-  padding: 0.5em;
-  line-height: 1;
-  background: #f6f6f7;
-  border: 0;
-  box-shadow: 0;
-  cursor: pointer;
-}
-
-._modal-close-icon {
-  width: 25px;
-  height: 25px;
-  fill: transparent;
-  stroke: black;
-  stroke-linecap: round;
-  stroke-width: 2;
-}
-
-.modal-body {
-  padding-top: 0.25em;
-}
-._hide-visual {
-  border: 0 !important;
-  clip: rect(0 0 0 0) !important;
-  height: 1px !important;
-  margin: -1px !important;
-  overflow: hidden !important;
-  padding: 0 !important;
-  position: absolute !important;
-  width: 1px !important;
-  white-space: nowrap !important;
-}
-
-.scroll-lock {
-  overflow: hidden;
-  margin-right: 17px;
 }
